pyproject.toml
[project]
name = "debaterag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-text-splitters>=1.1.0",
    "numpy>=2.4.2",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.30.0",
]
